Split ACQ-related constants to separate files based on domain
