+++ cypress/support/constants/consortia/index.js
@@ -1,0 +1,13 @@
+export const CONSORTIA_SYSTEM_USER = 'System, System user - mod-consortia-keycloak';
+
+export const PUBLISH_COORDINATOR_STATUSES = {
+  COMPLETE: 'COMPLETE',
+  ERROR: 'ERROR',
+  IN_PROGRESS: 'IN_PROGRESS',
+};
+
+export const PUBLISH_COORDINATOR_SHARE_DETAILS_KEYS = {
+  CREATE: 'createSettingsPCId',
+  DELETE: 'pcId',
+  UPDATE: 'updateSettingsPCId',
+};
