+++ cypress/support/constants/finance/expense-class.js
@@ -1,0 +1,10 @@
+export const EXPORT_EXPENSE_CLASS_FIELDS = {
+  EXPENSE_CLASS_NAME: 'Name (Exp Class)',
+  EXPENSE_CLASS_CODE: 'Code (Exp Class)',
+  EXPENSE_CLASS_STATUS: 'Status (Exp Class)',
+  EXPENSE_CLASS_ENCUMBERED: 'Encumbered (Exp Class)',
+  EXPENSE_CLASS_AWAITING_PAYMENT: 'Awaiting payment (Exp Class)',
+  EXPENSE_CLASS_EXPENDED: 'Expended (Exp Class)',
+  EXPENSE_CLASS_CREDITED: 'Credited (Exp Class)',
+  PERCENTAGE_OF_TOTAL_EXPENDED: 'Percentage of total expended',
+};
