+++ cypress/support/constants/finance/fiscal-year.js
@@ -1,0 +1,16 @@
+export const FUNDING_INFORMATION_NAMES = {
+  INITIAL_ALLOCATION: 'Initial allocation',
+  TOTAL_ALLOCATED: 'Total allocated',
+  TOTAL_FUNDING: 'Total funding',
+};
+
+export const FINANCIAL_ACTIVITY_OVERRAGES = {
+  ENCUMBERED: 'Encumbered',
+  AWAITING_PAYMENT: 'Awaiting payment',
+  EXPENDED: 'Expended',
+  CREDITED: 'Credited',
+  UNAVAILABLE: 'Unavailable',
+  OVER_ENCUMBRANCE: 'Over encumbrance',
+  OVER_EXPENDED: 'Over expended',
+  AVAILABLE_BALANCE: 'Available balance',
+};
