+++ cypress/support/constants/finance/fund.js
@@ -1,0 +1,17 @@
+export const FUND_DISTRIBUTION_TYPES = {
+  AMOUNT: 'amount',
+  PERCENTAGE: 'percentage',
+};
+
+export const EXPORT_FUND_FIELDS = {
+  FUND_NAME: 'Name (Fund)',
+  FUND_CODE: 'Code (Fund)',
+  FUND_STATUS: 'Status (Fund)',
+  FUND_TYPE: 'Type',
+  FUND_GROUPS: 'Group (Code)',
+  ACQ_UNITS: 'Acquisition unit',
+  TRANSFER_FROM: 'Transfer from',
+  TRANSFER_TO: 'Transfer to',
+  EXTERNAL_ACCOUNT_NO: 'External account number',
+  FUND_DESCRIPTION: 'Description',
+};
