+++ cypress/support/constants/finance/ledger.js
@@ -1,0 +1,12 @@
+export const LEDGER_VIEW_FIELDS = {
+  NAME: 'Name',
+  CODE: 'Code',
+  FISCAL_YEAR: 'Fiscal year',
+  STATUS: 'Status',
+};
+
+export const LEDGER_STATUSES = {
+  ACTIVE: 'Active',
+  FROZEN: 'Frozen',
+  INACTIVE: 'Inactive',
+};
