+++ cypress/support/constants/finance/transaction.js
@@ -1,0 +1,31 @@
+export const ENCUMBRANCE_STATUSES = {
+  RELEASED: 'Released',
+  UNRELEASED: 'Unreleased',
+};
+
+export const TRANSACTION_DETAIL_FIELDS = {
+  FISCAL_YEAR: 'Fiscal year',
+  TRANSACTION_DATE: 'Transaction date',
+  AMOUNT: 'Amount',
+  SOURCE: 'Source',
+  TYPE: 'Type',
+  FROM: 'From',
+  TO: 'To',
+  EXPENSE_CLASS: 'Expense class',
+  TAGS: 'Tags',
+  INITIAL_ENCUMBRANCE: 'Initial encumbrance',
+  AWAITING_PAYMENT: 'Awaiting payment',
+  EXPENDED: 'Expended',
+  STATUS: 'Status',
+  DESCRIPTION: 'Description',
+};
+
+export const TRANSACTION_TYPES = {
+  ALLOCATION: 'Allocation',
+  CREDIT: 'Credit',
+  ENCUMBRANCE: 'Encumbrance',
+  PAYMENT: 'Payment',
+  PENDING_PAYMENT: 'Pending payment',
+  ROLLOVER_TRANSFER: 'Rollover transfer',
+  TRANSFER: 'Transfer',
+};
