+++ cypress/support/constants/index.js
@@ -1,0 +1,8 @@
+/* TODO: reduce code duplications across constants */
+export * from './consortia';
+export * from './constants';
+export * from './finance';
+export * from './invoices';
+export * from './orders';
+export * from './organizations';
+export * from './receiving';
