+++ cypress/support/constants/invoices/invoice-line.js
@@ -1,0 +1,4 @@
+export const INVOICE_LINE_VIEW_FIELDS = {
+  STATUS: 'Status',
+  SUB_TOTAL: 'Sub-total',
+};
