+++ cypress/support/constants/invoices/invoice.js
@@ -1,0 +1,23 @@
+export const INVOICE_STATUSES = {
+  OPEN: 'Open',
+  REVIEWED: 'Reviewed',
+  APPROVED: 'Approved',
+  CANCELLED: 'Cancelled',
+  PAID: 'Paid',
+};
+
+export const INVOICE_POL_PAYMENT_STATUSES = {
+  AWAITING_PAYMENT: 'Awaiting Payment',
+  CANCELLED: 'Cancelled',
+  FULLY_PAID: 'Fully Paid',
+  NO_CHANGE: 'No Change',
+  PARTIALLY_PAID: 'Partially Paid',
+};
+
+export const INVOICE_VIEW_FIELDS = {
+  FISCAL_YEAR: 'Fiscal year',
+  INVOICE_STATUS: 'Status',
+  SUB_TOTAL: 'Sub-total',
+  VENDOR_NAME: 'Vendor name',
+  VOUCHER_STATUS: 'Status',
+};
