+++ cypress/support/constants/organizations/index.js
@@ -1,0 +1,5 @@
+export const ORGANIZATION_STATUSES = {
+  ACTIVE: 'Active',
+  INACTIVE: 'Inactive',
+  PENDING: 'Pending',
+};
