+++ cypress/support/constants/receiving/index.js
@@ -1,0 +1,9 @@
+export const RECEIVING_TITLE_ACCORDION_NAMES = {
+  TITLE_INFORMATION: 'Title information',
+  POL_DETAILS: 'POL details',
+  EXPECTED: 'Expected',
+  RECEIVED: 'Received',
+  ROUTING_LISTS: 'Routing lists',
+  UNRECEIVABLE: 'Unreceivable',
+  BOUND_ITEMS: 'Bound items',
+};
